pyproject.toml
[project]
name = "course-metadata-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.10"
dependencies = [
    "asyncpg>=0.30.0",
    "google-cloud-aiplatform>=1.122.0",
    "google-genai>=1.46.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.38.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0.1",
]

[tool.setuptools.package-dir]
assessment = "src/assessment"
